+++ packages/orq-rc/src/funcs/datasetsCreate.ts
@@ -101,7 +101,7 @@ async function $do(
   const context = {
     baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "CreateDataset",
-    oAuth2Scopes: null,
+    oAuth2Scopes: [],
 
     resolvedSecurity: requestSecurity,