+++ packages/orq-rc/src/funcs/knowledgeRetrieveDatasource.ts
@@ -105,7 +105,7 @@ async function $do(
   const context = {
     baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "RetrieveDatasource",
-    oAuth2Scopes: null,
+    oAuth2Scopes: [],
 
     resolvedSecurity: requestSecurity,