+++ packages/orq-rc/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.3.0-rc.13",
-  genVersion: "2.562.2",
-  userAgent: "speakeasy-sdk/typescript 3.3.0-rc.13 2.562.2 2.0 @orq-ai/node",
+  sdkVersion: "3.3.0-rc.14",
+  genVersion: "2.563.0",
+  userAgent: "speakeasy-sdk/typescript 3.3.0-rc.14 2.563.0 2.0 @orq-ai/node",
 } as const;