+++ packages/orq-rc/RELEASES.md
@@ -1109,3 +1109,13 @@ Based on:
 - [typescript v3.3.0-rc.13] packages/orq-rc
 ### Releases
 - [NPM v3.3.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.13 - packages/orq-rc
+
+## 2025-03-29 18:02:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.526.0 (2.563.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.0-rc.14] packages/orq-rc
+### Releases
+- [NPM v3.3.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.14 - packages/orq-rc